build: remove buildroot packages from download-info
Instead of keeping the list of archives in build/download-info which should be
downloaded for buildroot, we use "make source" to do the same job.
However, the task of building buildroot had to split into 3 tasks:
--init (unpack and configure buildroot)
--download (download the source of used buildroot packages)
--build (compile buildroot packages, etc.)

The reason for this change is to avoid updating the download-info file when
buildroot package is updated. Additionally ltrace is not downloaded as an
archive, but it is cloned git repo. After the clone it is packed into archive.
In the current state our framework cannot do the same.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>